apps/backend/requirements.txt
# FastAPI và web framework
fastapi==0.116.1
uvicorn==0.33.0
gunicorn==21.2.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.41
alembic==1.14.1
psycopg2-binary==2.9.10

# Configuration và validation
pydantic==2.10.6
pydantic-settings==2.3.4
python-dotenv==1.0.1
email-validator==2.1.0

# Security
python-jose[cryptography]==3.4.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0

# HTTP client
requests==2.31.0

# Background tasks
celery[redis]==5.3.4
redis>=4.5.2,<5.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code quality
flake8==6.1.0
black==23.12.1
isort==5.13.2

# Internationalization
babel==2.14.0

# Logging
structlog==23.2.0
